pyk/{ktool/kprove,kcfg/explore,integration_tests}: move RPC server handling to KCFGExplore from KProve
